Move serve into serve.go
